+++ frontend/tests/unit/components/Sign/SignTabitemsSignup.spec.js
